% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{add_tool}
\alias{add_tool}
\title{Add a tool to an agent}
\usage{
add_tool(x, tool, ...)
}
\arguments{
\item{x}{An object}

\item{tool}{A tool created with mcpr::new_tool}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified object
}
\description{
Add a tool to an agent
}
