% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{append_message}
\alias{append_message}
\title{Add a message to the list}
\usage{
append_message(x, message)
}
\arguments{
\item{message}{A message object.}

\item{provider}{An object of class \code{provider}.}
}
\description{
Add a message to the list
}
