% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{clear_messages}
\alias{clear_messages}
\title{Clear messages}
\usage{
clear_messages(provider)
}
\arguments{
\item{provider}{An object of class \code{provider}.}
}
\value{
A response object
}
\description{
Clear messages
}
\examples{
response <- clear_messages(provider)
}
