% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-ellmer.R
\name{convert_array_type}
\alias{convert_array_type}
\title{Convert array type to ellmer specification}
\usage{
convert_array_type(prop)
}
\arguments{
\item{prop}{Property definition}
}
\value{
ellmer type specification
}
\description{
Convert array type to ellmer specification
}
\keyword{internal}
