% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-ellmer.R
\name{convert_object_type}
\alias{convert_object_type}
\title{Convert object type to ellmer specification}
\usage{
convert_object_type(prop)
}
\arguments{
\item{prop}{Property definition}
}
\value{
ellmer type specification
}
\description{
Convert object type to ellmer specification
}
\keyword{internal}
