% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-ellmer.R
\name{convert_schema_to_ellmer_types}
\alias{convert_schema_to_ellmer_types}
\title{Convert JSON schema to ellmer type specifications}
\usage{
convert_schema_to_ellmer_types(schema)
}
\arguments{
\item{schema}{JSON schema object}
}
\value{
List of ellmer type specifications
}
\description{
Convert JSON schema to ellmer type specifications
}
\keyword{internal}
