% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_timediff}
\alias{format_timediff}
\title{Format a difftime object as a string}
\usage{
format_timediff(td, precision = 3)
}
\arguments{
\item{td}{A difftime object.}

\item{precision}{The number of significant digits to include in the output.}
}
\value{
A string representation of the difftime object.
}
\description{
Formats a difftime object as a string with a specified precision.
}
