% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{get_last_message}
\alias{get_last_message}
\title{Get last message}
\usage{
get_last_message(x)
}
\arguments{
\item{x}{An object of class \code{agent}.}
}
\value{
A list of messages
}
\description{
Get last message
}
