% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_session_data}
\alias{get_session_data}
\title{Get LLM session data}
\usage{
get_session_data(id = NULL)
}
\arguments{
\item{id}{The session ID to retrieve the data for. If not provided, all
session data will be returned.}
}
\value{
A list with the session data or a list of lists of session data.
}
\description{
Retrieve the data stored in the session option for a specific session ID.
}
