% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_session_data_ids}
\alias{get_session_data_ids}
\title{Get LLM session data IDs}
\usage{
get_session_data_ids()
}
\value{
A character vector with the session IDs.
}
\description{
Retrieve the session IDs for the stored session data.
}
