% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-ellmer.R
\name{handle_response.provider_ellmer}
\alias{handle_response.provider_ellmer}
\title{Handle response for ellmer providers}
\usage{
\method{handle_response}{provider_ellmer}(x, agent, response, loop = TRUE)
}
\arguments{
\item{x}{An ellmer provider object}

\item{agent}{An agent object}

\item{response}{Response from ellmer}

\item{loop}{Whether to continue tool loops}
}
\value{
Updated agent
}
\description{
Handle response for ellmer providers
}
