% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{handle_response}
\alias{handle_response}
\title{Handle response from an LLM provider}
\usage{
handle_response(provider, response)
}
\arguments{
\item{provider}{An object of class \code{provider}.}

\item{response}{A response object from the LLM provider.}
}
\value{
A response object
}
\description{
Handle response from an LLM provider
}
