% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{handle_tool_use}
\alias{handle_tool_use}
\title{Handle tool use in a response}
\usage{
handle_tool_use(provider, response)
}
\arguments{
\item{provider}{An object of class \code{provider}.}

\item{response}{A response object from the LLM provider.}
}
\value{
A formatted tool response
}
\description{
Handle tool use in a response
}
