% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_management.R
\name{is_rate_limit_exceeded}
\alias{is_rate_limit_exceeded}
\title{Check if the query rate limit has been exceeded}
\usage{
is_rate_limit_exceeded(
  response,
  wait = TRUE,
  max_attempts = getOption("llmr_retry_attempts", 3),
  log_request = getOption("llmr_log_request", FALSE)
)
}
\arguments{
\item{response}{The response object from an LLM calling function.}

\item{wait}{If \code{TRUE}, the function will wait for the time specified in the
response error message before retrying the request.}

\item{max_attempts}{The number of times to retry the request before rising an
error and stopping the process.}

\item{log_request}{If \code{TRUE}, the error message from the response will be
printed to the console.}
}
\value{
\code{TRUE} if the request failed due to rate limiting
}
\description{
Check if the query rate limit for the LLM provider has been exceeded. If the
rate limit has been exceeded, the function will wait for the specified time
before retrying the request.
}
