% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-ellmer.R
\name{mcpr_to_ellmer_tool}
\alias{mcpr_to_ellmer_tool}
\title{Convert mcpr tool to ellmer ToolDef}
\usage{
mcpr_to_ellmer_tool(mcpr_tool)
}
\arguments{
\item{mcpr_tool}{A tool object from mcpr package}
}
\value{
An ellmer ToolDef object
}
\description{
Convert mcpr tool to ellmer ToolDef
}
\keyword{internal}
