% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message.R
\name{message}
\alias{message}
\alias{new_message}
\title{Create a message}
\usage{
new_message(content, role = "user")
}
\arguments{
\item{content}{Content of the message}

\item{role}{Role of the message}
}
\value{
A message object
}
\description{
Create a message
}
\examples{
message <- new_message("Hello, world!")
}
