% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{new_mcp}
\alias{new_mcp}
\title{Create a new mcp provider}
\usage{
new_mcp(command, args = character(), name = command, version = "1.0.0")
}
\arguments{
\item{command}{The command to run}

\item{args}{Arguments to pass to the command}

\item{name}{The name of the provider}
}
\value{
A new mcp provider
}
\description{
Create a new mcp provider
}
