% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_prompting.R
\name{process_messages}
\alias{process_messages}
\title{Process chat message into standard format}
\usage{
process_messages(messages)
}
\arguments{
\item{messages}{A character vector or list of chat messages. It can be a
vector or a specifically structured list.}
}
\value{
A list of chat messages in standard format.
}
\description{
This function takes one or more chat messages and processes them into a
standard list format with role and content for each message to be fed to a
large language model.
}
\details{
The standard format is a list of chat messages with the following structure:
message: \code{c(role = "system", content = "Welcome to the chat!")}

list of messages: \code{list(
   c(role = "system", content = "You are an useful AI assistant."),
   c(role = "user", content = "Hi there!")
 )}

list format: \code{list(
   list(role = "system", content = "You are an useful AI assistant."),
   list(role = "user", content = "Hi there!")
 )}
}
\examples{
process_messages(c(user = "Hello, AI!"))
process_messages(list(list(role = "user", content = "How are you?")))

}
