% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider.R
\name{provider}
\alias{provider}
\alias{new_provider}
\alias{new_anthropic}
\alias{new_openai}
\title{Create a provider}
\usage{
new_provider(name, url, ...)

new_anthropic(url = "https://api.anthropic.com", ...)

new_openai(url = "https://api.openai.com", ...)
}
\arguments{
\item{name}{Name of the provider}

\item{url}{URL of the provider}

\item{...}{Additional options}
}
\value{
A provider object
}
\description{
Create a provider
}
\examples{
provider <- new_anthropic()
}
