% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-jsonrpc.R
\name{read}
\alias{read}
\title{Read a JSON-RPC response from a mcp provider}
\usage{
read(x)
}
\arguments{
\item{x}{A mcp provider}
}
\value{
The response
}
\description{
Read a JSON-RPC response from a mcp provider
}
