% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{record_llmr_model}
\alias{record_llmr_model}
\title{Store LLM model specifications}
\usage{
record_llmr_model(
  provider,
  api_type = c("custom", "openai", "azure", "gemini"),
  endpoint = NULL,
  api_key = NULL,
  model = NULL,
  api_version = NULL
)
}
\arguments{
\item{provider}{The provider name under which to store the model
specifications.}

\item{api_type}{The type of API used by the provider (custom, openai, azure,
or gemini).}

\item{endpoint}{The API endpoint to use for the provider.}

\item{api_key}{The API key to use for the provider.}

\item{model}{The default model to use for the provider. Some providers may
accept only one model, while others do not require a model to be specified.}

\item{api_version}{The version of the API to use (only for "azure" api types
for now)}
}
\description{
Store the specifications for a language model in the global options.
}
