% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{register_agent}
\alias{register_agent}
\title{Register an agent with a provider}
\usage{
register_agent(provider, agent, ...)
}
\arguments{
\item{provider}{An object of class \code{provider}}

\item{agent}{An object of class \code{agent}}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified provider object (invisibly)
}
\description{
Register an agent with a provider
}
