% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request-ellmer.R
\name{request.provider_ellmer}
\alias{request.provider_ellmer}
\title{Request method for ellmer providers}
\usage{
\method{request}{provider_ellmer}(x, message = NULL, ..., tools = NULL)
}
\arguments{
\item{x}{An ellmer provider object}

\item{message}{A message or list of messages}

\item{...}{Additional arguments}

\item{tools}{Optional list of tools}
}
\value{
Response from ellmer
}
\description{
Request method for ellmer providers
}
