% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Make a request to an LLM provider}
\usage{
request(x, message = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{provider} or \code{agent}.}

\item{message}{A message object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A response object
}
\description{
Make a request to an LLM provider
}
\details{
You can make a request with \code{message = NULL}
The agent will send the history of the conversation to the LLM provider.
}
