% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the key for an LLM provider}
\usage{
set_api_key(x, key, ...)
}
\arguments{
\item{x}{A provider object}

\item{key}{Character string specifying the API key}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified object
}
\description{
Set the key for an LLM provider
}
