% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ellmer.R
\name{set_max_tokens.provider_ellmer}
\alias{set_max_tokens.provider_ellmer}
\title{Set max tokens for ellmer provider}
\usage{
\method{set_max_tokens}{provider_ellmer}(x, max, ...)
}
\arguments{
\item{x}{An ellmer provider object}

\item{max}{Maximum tokens}

\item{...}{Additional arguments}
}
\value{
The modified provider
}
\description{
Set max tokens for ellmer provider
}
