% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-tokens.R
\name{set_max_tokens}
\alias{set_max_tokens}
\title{Set the token for an LLM provider}
\usage{
set_max_tokens(x, max, ...)
}
\arguments{
\item{x}{A provider object}

\item{max}{Integer specifying the maximum number of tokens to generate}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified object
}
\description{
Set the token for an LLM provider
}
