% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_max_tool_calls}
\alias{set_max_tool_calls}
\title{Set the maximum number of tool calls for an LLM provider}
\usage{
set_max_tool_calls(x, max, ...)
}
\arguments{
\item{x}{An object}

\item{max}{Character string containing the system max}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified object
}
\description{
Set the maximum number of tool calls for an LLM provider
}
