% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ellmer.R
\name{set_model.provider_ellmer}
\alias{set_model.provider_ellmer}
\title{Set model for ellmer provider}
\usage{
\method{set_model}{provider_ellmer}(x, name, ...)
}
\arguments{
\item{x}{An ellmer provider object}

\item{name}{Model name}

\item{...}{Additional arguments}
}
\value{
The modified provider
}
\description{
Set model for ellmer provider
}
