% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_session_id}
\alias{set_session_id}
\title{Set the LLM session ID}
\usage{
set_session_id(session_id = NULL)
}
\arguments{
\item{session_id}{The session ID to store the data under. If not provided, a
new one will be generated. Usually a number prepend with a hash (#).}
}
\value{
The session ID, invisibly.
}
\description{
Set the session ID to store the data from the interaction with a language
model.
}
