% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ellmer.R
\name{set_temperature.provider_ellmer}
\alias{set_temperature.provider_ellmer}
\title{Set temperature for ellmer provider}
\usage{
\method{set_temperature}{provider_ellmer}(x, temperature, ...)
}
\arguments{
\item{x}{An ellmer provider object}

\item{temperature}{Temperature value}

\item{...}{Additional arguments}
}
\value{
The modified provider
}
\description{
Set temperature for ellmer provider
}
