% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-ellmer.R
\name{set_version.provider_ellmer}
\alias{set_version.provider_ellmer}
\title{Set version for ellmer provider}
\usage{
\method{set_version}{provider_ellmer}(x, version, ...)
}
\arguments{
\item{x}{An ellmer provider object}

\item{version}{API version}

\item{...}{Additional arguments}
}
\value{
The modified provider
}
\description{
Set version for ellmer provider
}
