% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{step}
\alias{step}
\title{Create a workflow step}
\usage{
step(x, ...)
}
\arguments{
\item{x}{An agent or function to use as a step}

\item{...}{Additional arguments}

\item{name}{Optional name for the step}
}
\value{
A workflow step object
}
\description{
Creates a step that can be used in workflows. Steps can wrap functions
or agents to provide processing capabilities.
}
