% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_management.R
\name{stop_on_no_output}
\alias{stop_on_no_output}
\title{Stop if no tokens were generated}
\usage{
stop_on_no_output(parsed)
}
\arguments{
\item{parsed}{The parsed response object from an LLM calling function.}
}
\description{
Stop the execution of the script if no tokens were generated by the LLM
provider.
}
