% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_management.R
\name{stop_on_response_error}
\alias{stop_on_response_error}
\title{Stop on response error}
\usage{
stop_on_response_error(response)
}
\arguments{
\item{response}{The response object from an LLM calling function.}
}
\description{
Stop the execution of the script if the response object from an LLM calling
function contains an error message.
}
