% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{store_llm_session_data}
\alias{store_llm_session_data}
\title{Store LLM session data}
\usage{
store_llm_session_data(
  messages,
  parameters = NULL,
  response,
  usage = NULL,
  processing_time = NULL,
  provider = getOption("llmr_provider", NULL),
  model = getOption("llmr_model", NULL),
  session_id = getOption("llmr_session_id", NULL)
)
}
\arguments{
\item{messages}{A list of messages sent to the language model.}

\item{parameters}{A list of parameters used in the request.}

\item{response}{The response from the language model as a string or a vector.}

\item{usage}{The token usage information from the response.}

\item{processing_time}{The time taken to process the request.}

\item{provider}{The language model provider.}

\item{model}{The language model used.}

\item{session_id}{The LLM session ID to store the data under. Can be used to continue a session.}
}
\description{
Store information about the interaction with a language model in a session
option.
}
