% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{sync_ellmer_history}
\alias{sync_ellmer_history}
\title{Sync conversation history with ellmer Chat object}
\usage{
sync_ellmer_history(x)
}
\arguments{
\item{x}{An agent with ellmer Chat provider}
}
\description{
Sync conversation history with ellmer Chat object
}
\keyword{internal}
