% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_prompting.R
\name{use_gemini_llm}
\alias{use_gemini_llm}
\title{Use Google Gemini Language Model}
\usage{
use_gemini_llm(
  body,
  model = getOption("llmr_model"),
  api_key = getOption("llmr_api_key"),
  log_request = getOption("llmr_log_requests", TRUE)
)
}
\arguments{
\item{body}{The body of the request.}

\item{model}{Model identifier for the Google Gemini API. Obtained from R
options.}

\item{api_key}{API key for the Google Gemini service. Obtained from R
options.}

\item{log_request}{A boolean to log the request time. Can be set up globally
using the \code{llmr_log_requests} option, which defaults to TRUE.}
}
\value{
The function returns the response from the Google Gemini API.
}
\description{
Sends a request to the Google Gemini API using the parameters in the \code{body}
argument. It requires an API key set in the R options. Internally, the body
and the output are transformed to be compatible with the OpenAI API. Users
would not typically call this function directly, but rather use the
\code{prompt_llm} function.
}
