% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_prompting.R
\name{use_openai_llm}
\alias{use_openai_llm}
\title{Use OpenAI Language Model}
\usage{
use_openai_llm(
  body,
  model = getOption("llmr_openai_model_gpt"),
  api_key = getOption("llmr_openai_api_key"),
  log_request = getOption("llmr_log_requests", TRUE)
)
}
\arguments{
\item{body}{The body of the request.}

\item{model}{Model identifier for the OpenAI API. Obtained from R options.}

\item{api_key}{API key for the OpenAI service. Obtained from R options.}

\item{log_request}{A boolean to log the request time. Can be set up globally
using the \code{llmr_log_requests} option, which defaults to TRUE.}
}
\value{
The function returns the response from the OpenAI API.
}
\description{
Sends a request to the OpenAI API using the parameters in the \code{body}
argument. It requires an API key and model identifier set in the R options.
}
