% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp-jsonrpc.R
\name{write}
\alias{write}
\title{Write a JSON-RPC request to a mcp provider}
\usage{
write(x, method, params = NULL, id = generate_id())
}
\arguments{
\item{x}{A mcp provider}

\item{method}{The method to call}

\item{params}{The parameters to pass to the method}

\item{id}{The id of the request}
}
\value{
The mcp provider
}
\description{
Write a JSON-RPC request to a mcp provider
}
