% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{find_mcp_by_name}
\alias{find_mcp_by_name}
\title{Find an MCP by name}
\usage{
find_mcp_by_name(env, name)
}
\arguments{
\item{env}{An environment containing MCPs}

\item{name}{The name of the MCP to find.}
}
\value{
An MCP object or NULL if not found
}
\description{
Find an MCP by name
}
