% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_session_data}
\alias{get_session_data}
\title{Get LLM session data}
\usage{
get_session_data(id = NULL)
}
\arguments{
\item{id}{The session ID to retrieve the data for. If not provided, all
session data will be returned.}
}
\value{
A list with the session data for the specified ID, or a list of lists
containing all session data if no ID is provided. Each session data entry
contains information about the messages, parameters, response, token usage,
processing time, provider, and model used in the interaction.
}
\description{
Retrieve the data stored in the session option for a specific session ID.
}
\examples{
# Get data for a specific session
session_data <- get_session_data("#12345")

# Get all session data
all_data <- get_session_data()

}
