% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{new_agent}
\alias{new_agent}
\title{Create a new agent}
\usage{
new_agent(name)
}
\arguments{
\item{name}{Character string specifying the agent name}
}
\value{
An object of class "agent"
}
\description{
Creates a new agent with the specified name
}
