% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{record_llmr_model}
\alias{record_llmr_model}
\title{Store LLM model specifications}
\usage{
record_llmr_model(
  label,
  provider = c("custom", "openai", "azure", "gemini"),
  endpoint = NULL,
  api_key = NULL,
  model = NULL,
  api_version = NULL,
  parameters = NULL
)
}
\arguments{
\item{label}{A label under which to store the model specification.}

\item{provider}{The provider of the API (custom, openai, azure,
or gemini).}

\item{endpoint}{The API endpoint to use for the provider.}

\item{api_key}{The API key to use for the provider.}

\item{model}{The default model to use for the provider. Some providers may
accept only one model, while others do not require a model to be specified.}

\item{api_version}{The version of the API to use (only for "azure" api types
for now)}

\item{parameters}{A list of default parameters to use with this model. These
parameters will be used as defaults in prompt_llm() and can be overridden
by parameters passed directly to prompt_llm().}
}
\description{
Store the specifications for a language model in the global options.
}
