% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp.R
\name{register_mcp}
\alias{register_mcp}
\title{Register a memory client provider}
\usage{
register_mcp(x, mcp)
}
\arguments{
\item{x}{An object of class \code{agent}.}

\item{mcp}{An object of class \code{client} from the "mcpr" package.}
}
\value{
A provider object
}
\description{
Register a memory client provider
}
