% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request}
\alias{request}
\title{Make a request to an LLM provider}
\usage{
request(provider, message)
}
\arguments{
\item{provider}{An object of class \code{provider}.}

\item{message}{A message object.}
}
\value{
A response object
}
\description{
Make a request to an LLM provider
}
\examples{
response <- request(provider, endpoint)
}
