% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_llmr_model}
\alias{set_llmr_model}
\title{Set the current LLMR model}
\usage{
set_llmr_model(label, model = NULL)
}
\arguments{
\item{label}{The label to choose a model among the stored ones.}

\item{model}{An optional model to override the default model for the
provider, if any.}
}
\value{
Invisibly, the updated global options.
}
\description{
This function allows you to set the current LLMR model to use for subsequent
operations. It retrieves the stored model specifications using the label
under which it is stored and updates the global options accordingly.
}
