% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{set_temperature}
\alias{set_temperature}
\title{Set the temperature for an LLM provider}
\usage{
set_temperature(x, temperature, ...)
}
\arguments{
\item{x}{A provider object}

\item{temperature}{Numeric value specifying the temperature (typically between 0 and 1)}

\item{...}{Additional arguments passed to methods}
}
\value{
The modified object
}
\description{
Set the temperature for an LLM provider
}
